+++ ProjectManagement/ProjectManagement.Core/UseCases/Projects/Dto/DetailedProjectDto.cs
@@ -1,23 +1,23 @@
 ﻿using Domain.Entities;
-using Newtonsoft.Json;
 using ProjectManagement.Core.Base.Mappings;
 using System;
 
-namespace ProjectManagement.Core.UseCases.Projects.Queries.GetProjects.Dto
+namespace ProjectManagement.Core.UseCases.Projects.Dto
 {
-    public class ProjectDto : IMapFrom<Project>
+    public class DetailedProjectDto : IMapFrom<Project>
     {
-        [JsonProperty(Order = int.MinValue)]
         public int Id { get; set; }
         public string Name { get; set; }
+        public string Description { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime TargetDate { get; set; }
         public DateTime EndDate { get; set; }
         public bool IsActive { get; set; }
+        public int CustomerId { get; set; }
 
         public static void Mapping(MappingProfile profile)
         {
-            profile.CreateMap<Project, ProjectDto>();
+            profile.CreateMap<Project, DetailedProjectDto>();
         }
     }
 }
