+++ ProjectManagement/ProjectManagement.Core/UseCases/Projects/Queries/GetProjectById/GetProjectByIdQueryHandler.cs
@@ -3,7 +3,7 @@ using Domain.Entities;
 using MediatR;
 using ProjectManagement.Core.Base.Exceptions;
 using ProjectManagement.Core.Base.Interfaces;
-using ProjectManagement.Core.UseCases.Projects.Queries.GetProjectById.Dto;
+using ProjectManagement.Core.UseCases.Projects.Dto;
 using System.Threading;
 using System.Threading.Tasks;
