+++ ProjectManagement/ProjectManagement.Core/UseCases/Projects/Queries/GetProjects/GetProjectsQueryHandler.cs
@@ -3,7 +3,7 @@ using AutoMapper.QueryableExtensions;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
 using ProjectManagement.Core.Base.Interfaces;
-using ProjectManagement.Core.UseCases.Projects.Queries.GetProjects.Dto;
+using ProjectManagement.Core.UseCases.Projects.Dto;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
