+++ ProjectManagement/ProjectManagement.Core/UseCases/Projects/Queries/GetProjects/ProjectVm.cs
@@ -1,4 +1,4 @@
-﻿using ProjectManagement.Core.UseCases.Projects.Queries.GetProjects.Dto;
+﻿using ProjectManagement.Core.UseCases.Projects.Dto;
 using System.Collections.Generic;
 
 namespace ProjectManagement.Core.UseCases.Projects.Queries.GetProjects
