+++ ProjectManagement/ProjectManagementApi/Controllers/ProjectController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ProjectManagement.Core.UseCases.Projects.Commands.CreateProject;
 using ProjectManagement.Core.UseCases.Projects.Commands.DeleteProject;
+using ProjectManagement.Core.UseCases.Projects.Dto;
 using ProjectManagement.Core.UseCases.Projects.Queries.GetProjectById;
-using ProjectManagement.Core.UseCases.Projects.Queries.GetProjectById.Dto;
 using ProjectManagement.Core.UseCases.Projects.Queries.GetProjects;
 using System.Threading.Tasks;
